Fixed sounds in nodes
The sounds in node definitions were using a function from the default mod, but wasn't calling it, resulting in a game stopping error. The engine expected a table and got a function, so I added parenthesis to each sound so that a table is returned and used for the sound as expected.